+++ Generics/Program.cs
@@ -12,12 +12,12 @@ namespace Generics
         {
             Maximum maximum = new Maximum();
             Console.WriteLine("Enter the FirstNumber");
-            int firstvalue = Convert.ToInt32(Console.ReadLine());
+            float firstvalue = Convert.ToSingle(Console.ReadLine());
             Console.WriteLine("Enter the SecondNumber");
-            int secondvalue = Convert.ToInt32(Console.ReadLine());
+            float secondvalue = Convert.ToSingle(Console.ReadLine());
             Console.WriteLine("Enter the ThirdNumber");
-            int thirdvalue = Convert.ToInt32(Console.ReadLine());
-            int val = maximum.Number(firstvalue, secondvalue, thirdvalue);
+            float thirdvalue = Convert.ToSingle(Console.ReadLine());
+            float val = maximum.Number(firstvalue, secondvalue, thirdvalue);
             Console.WriteLine("{0} is largest number ", val);
             Console.ReadLine();
         }
